Fix ModuleList bug
